Correctly split out manifest service into separate source files
